Teraform/cloud-functions/AddReviews/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-language